backendApi/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.32.1
mysql-connector-python==9.2.0
pydantic==2.7.1
pymongo==4.13.2
requests==2.32.2
pyarrow==19.0.1
pandas==2.2.2
fastapi-pagination==0.12.25
duckdb==1.0.0
python-dotenv==1.0.1
docker==7.1.0
elasticsearch==8.15.1
detect_delimiter==0.1.1
xlrd==2.0.1
python-magic==0.4.27
dict-path-finder==0.0.3
python-benedict==0.33.2
sqlalchemy==2.0.36
pytest==8.3.3
oauthlib==3.2.2
openpyxl==3.1.5
fastavro==1.10.0
beanie==1.30.0
motor==3.7.1
pydantic-settings==2.1.0
pymongoarrow==1.7.1
pytest-asyncio==1.0.0
mongomock-motor==0.0.36
python-multipart==0.0.17
httpx==0.27.2
gunicorn==21.2.0
jsonschema==4.24.0